move all image-related util methods to MediaUtil, refactor CashingAsyncImageLoader to use inflateDisplayImage() method
