+++ app/src/org/odk/collect/android/widgets/GridMultiWidget.java
@@ -18,6 +18,7 @@ import android.widget.GridView;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import org.commcare.android.util.MediaUtil;
 import org.commcare.android.util.StringUtils;
 import org.commcare.dalvik.R;
 import org.javarosa.core.model.SelectChoice;
@@ -28,7 +29,6 @@ import org.javarosa.core.reference.InvalidReferenceException;
 import org.javarosa.core.reference.ReferenceManager;
 import org.javarosa.form.api.FormEntryCaption;
 import org.javarosa.form.api.FormEntryPrompt;
-import org.odk.collect.android.utilities.FileUtils;
 
 import java.io.File;
 import java.util.Vector;
@@ -107,9 +107,8 @@ public class GridMultiWidget extends QuestionWidget {
                                     .getDefaultDisplay();
                         int screenWidth = display.getWidth();
                         int screenHeight = display.getHeight();
-                        Bitmap b =
-                            FileUtils
-                                    .getBitmapScaledToContainer(imageFile, screenHeight, screenWidth);
+                        Bitmap b = MediaUtil
+                                .getBitmapScaledToContainer(imageFile, screenHeight, screenWidth);
                         if (b != null) {
 
                             if (b.getWidth() > maxColumnWidth) {
@@ -292,9 +291,8 @@ public class GridMultiWidget extends QuestionWidget {
                                     .getDefaultDisplay();
                         int screenWidth = display.getWidth();
                         int screenHeight = display.getHeight();
-                        Bitmap b =
-                            FileUtils
-                                    .getBitmapScaledToContainer(imageFile, screenHeight, screenWidth);
+                        Bitmap b = MediaUtil
+                                .getBitmapScaledToContainer(imageFile, screenHeight, screenWidth);
                         if (b != null) {
 
                             if (mImageView == null) {
