+++ app/src/org/odk/collect/android/widgets/GridWidget.java
@@ -18,6 +18,7 @@ import android.widget.GridView;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import org.commcare.android.util.MediaUtil;
 import org.commcare.android.util.StringUtils;
 import org.commcare.dalvik.R;
 import org.javarosa.core.model.SelectChoice;
@@ -30,7 +31,6 @@ import org.javarosa.form.api.FormEntryCaption;
 import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.listeners.AdvanceToNextListener;
 import org.odk.collect.android.listeners.WidgetChangedListener;
-import org.odk.collect.android.utilities.FileUtils;
 
 import java.io.File;
 import java.util.Vector;
@@ -116,7 +116,7 @@ public class GridWidget extends QuestionWidget {
                         int screenWidth = display.getWidth();
                         int screenHeight = display.getHeight();
                         Bitmap b =
-                            FileUtils
+                                MediaUtil
                                     .getBitmapScaledToContainer(imageFile, screenHeight, screenWidth);
                         if (b != null) {
 
@@ -246,7 +246,7 @@ public class GridWidget extends QuestionWidget {
                         int screenWidth = display.getWidth();
                         int screenHeight = display.getHeight();
                         Bitmap b =
-                            FileUtils
+                                MediaUtil
                                     .getBitmapScaledToContainer(imageFile, screenHeight, screenWidth);
                         if (b != null) {
 
@@ -418,7 +418,7 @@ public class GridWidget extends QuestionWidget {
                         int screenWidth = display.getWidth();
                         int screenHeight = display.getHeight();
                         Bitmap b =
-                            FileUtils
+                                MediaUtil
                                     .getBitmapScaledToContainer(imageFile, screenHeight, screenWidth);
                         if (b != null) {
