+++ app/src/org/odk/collect/android/widgets/SignatureWidget.java
@@ -32,6 +32,7 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import org.commcare.android.util.MediaUtil;
 import org.commcare.android.util.StringUtils;
 import org.commcare.dalvik.R;
 import org.javarosa.core.model.data.IAnswerData;
@@ -40,7 +41,6 @@ import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.activities.DrawActivity;
 import org.odk.collect.android.activities.FormEntryActivity;
 import org.odk.collect.android.application.Collect;
-import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.android.utilities.UrlUtils;
 
 import java.io.File;
@@ -115,7 +115,7 @@ public class SignatureWidget extends QuestionWidget implements IBinaryWidget {
             File f = new File(mInstanceFolder + File.separator + mBinaryName);
 
             if (f.exists()) {
-                Bitmap bmp = FileUtils.getBitmapScaledToContainer(f, screenHeight, screenWidth);
+                Bitmap bmp = MediaUtil.getBitmapScaledToContainer(f, screenHeight, screenWidth);
                 if (bmp == null) {
                     mErrorTextView.setVisibility(View.VISIBLE);
                 }
